Split initialization from gyro.c for flash savings Move low performance requirements initialization code into gyro_init.c and optimize that for size.
