+++ src/main/cli/cli.c
@@ -164,6 +164,7 @@ bool cliMode = false;
 #include "sensors/compass.h"
 #include "sensors/esc_sensor.h"
 #include "sensors/gyro.h"
+#include "sensors/gyro_init.h"
 #include "sensors/sensors.h"
 
 #include "telemetry/frsky_hub.h"
